Revert "Add files via upload"
