--- Makefile.win
@@ -1,40 +1,0 @@
-# Project: 项目3
-# Makefile created by Dev-C++ 5.11
-
-CPP      = g++.exe -D__DEBUG__
-CC       = gcc.exe -D__DEBUG__
-WINDRES  = windres.exe
-OBJ      = main.o rect.o circle.o triangle.o line.o
-LINKOBJ  = main.o rect.o circle.o triangle.o line.o
-LIBS     = -L"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib32" -L"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/lib32" -static-libgcc -lglut32  -lglu32  -lopengl32  -lwinmm  -lgdi32 glut.dll glut32.dll -m32 -g3
-INCS     = -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include"
-CXXINCS  = -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++"
-BIN      = 项目3.exe
-CXXFLAGS = $(CXXINCS) -m32 -g3
-CFLAGS   = $(INCS) -m32 -g3
-RM       = rm.exe -f
-
-.PHONY: all all-before all-after clean clean-custom
-
-all: all-before $(BIN) all-after
-
-clean: clean-custom
-	${RM} $(OBJ) $(BIN)
-
-$(BIN): $(OBJ)
-	$(CPP) $(LINKOBJ) -o $(BIN) $(LIBS)
-
-main.o: main.cpp
-	$(CPP) -c main.cpp -o main.o $(CXXFLAGS)
-
-rect.o: rect.cpp
-	$(CPP) -c rect.cpp -o rect.o $(CXXFLAGS)
-
-circle.o: circle.cpp
-	$(CPP) -c circle.cpp -o circle.o $(CXXFLAGS)
-
-triangle.o: triangle.cpp
-	$(CPP) -c triangle.cpp -o triangle.o $(CXXFLAGS)
-
-line.o: line.cpp
-	$(CPP) -c line.cpp -o line.o $(CXXFLAGS)
