--- Readme.md
@@ -1,3 +1,0 @@
-## 计算机图形学实验
-
-基于openGL的demo
