--- func.h
@@ -1,14 +1,0 @@
-#ifndef FUNCTIONS_H
-#define FUNCTIONS_H
-#define GLUT_DISABLE_ATEXIT_HACK
-#include<windows.h>
-
-#include "glut.h" 
-#include<iostream>
-#include<cmath>
-#include<math.h>
-void drawRect(int argc, char* argv[]);
-void displayCircle(int argc, char* argv[]); 
-void drawTri(int argc, char * argv[]);
-void drawLine(int argc, char * argv[]);
-#endif // FUNCTIONS_H
