--- 项目3.layout
@@ -1,43 +1,0 @@
-[Editor_1]
-CursorCol=1
-CursorRow=19
-TopLine=1
-LeftChar=1
-[Editors]
-Order=0,2,3,4,1,5,6,7
-Focused=7
-[Editor_0]
-CursorCol=10
-CursorRow=13
-TopLine=1
-LeftChar=1
-[Editor_2]
-CursorCol=39
-CursorRow=15
-TopLine=1
-LeftChar=1
-[Editor_3]
-CursorCol=27
-CursorRow=167
-TopLine=1
-LeftChar=1
-[Editor_4]
-CursorCol=40
-CursorRow=13
-TopLine=1
-LeftChar=1
-[Editor_5]
-CursorCol=22
-CursorRow=36
-TopLine=25
-LeftChar=1
-[Editor_6]
-CursorCol=29
-CursorRow=208
-TopLine=194
-LeftChar=1
-[Editor_7]
-CursorCol=17
-CursorRow=3
-TopLine=1
-LeftChar=1
